requirements.txt
flask==2.2.3
python-pptx==0.6.21
werkzeug==2.2.3
requests==2.28.2
azure-identity==1.12.0
azure-storage-blob==12.14.1
pillow==9.4.0
moviepy==2.1.2